+++ backend-ai-webui-react/src/index.tsx
@@ -1,13 +1,27 @@
+import DefaultProviders from "./components/DefaultProviders";
 import ExampleComponent from "./components/ExampleComponent";
 import ResourceMonitor from "./components/ResourceMonitor";
-import { reactToWebComponentWithDefault } from "./helper";
+import reactToWebComponent from "./helper/react-to-webcomponent";
+
+// customElements.define(
+//   "backend-ai-webui-react-example",
+//   reactToWebComponentWithDefault(ExampleComponent)
+// );
 
 customElements.define(
   "backend-ai-webui-react-example",
-  reactToWebComponentWithDefault(ExampleComponent)
+  reactToWebComponent((props) => (
+    <DefaultProviders {...props}>
+      <ExampleComponent />
+    </DefaultProviders>
+  ))
 );
+
 customElements.define(
   "backend-ai-webui-react-resource-monitor",
-  //@ts-ignore
-  reactToWebComponentWithDefault(ResourceMonitor)
+  reactToWebComponent((props) => (
+    <DefaultProviders {...props}>
+      <ResourceMonitor />
+    </DefaultProviders>
+  ))
 );
